project

+++ emc-task-impl/src/main/java/com/huak/weather/WeatherServiceImpl.java
@@ -60,8 +60,11 @@ public class WeatherServiceImpl implements WeatherTaskService{
     public void executeWeather7dTask(List<Map<String, Object>> params) {
         if(params != null){
             for (Map param : params){
-                weekforcastDao.deletebyParmas(param);
-                this.weekForcast(param);
+                List<Weekforcast> tempLists = weekforcastDao.selectByParams(param);
+                if(tempLists != null && tempLists.size()>0){
+                    weekforcastDao.deletebyParmas(param);
+                }
+                 this.weekForcast(param);
             }
         }
     }
